+++ src/monitoring/ProgramMetrics.cs
@@ -1,0 +1,4 @@
+public static class ProgramMetrics
+{
+    public static MetricsCollector Instance { get; } = new MetricsCollector();
+}