+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 Všechny významné změny tohoto projektu budou zaznamenány v tomto souboru.
 Formát vychází z [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## 2026-01-24 - Martin Šilar
+
+### Přidáno
+
+- Monitoring statistik node (snapshoty + command metrics ukládané do MySQL)
+
+---
+
 ## 2026-01-23 - Martin Šilar
 
 ### Přidáno